Move IsConvertible and GetAstIdentifierForOutputIndex to base class
